requirements.txt
python-dotenv==0.19.2
requests==2.27.1
PyYAML==6.0
mysql-connector-python==8.0.29
dnspython==2.2.1
num2words==0.5.12
discord.py==2.2.2
discord.py-Console==0.1.2
lavalink==4.0.3
Pillow==9.3.0
pyarr==3.1.3
openai==0.27.1